feat: rewrites tests after refactoring
