+++ src/main/java/at/fh/burgenland/HelloApplication.java
@@ -6,11 +6,14 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-/** JavaFX example using for pipeline test. */
+/**
+ * JavaFX example using for pipeline test.
+ */
 public class HelloApplication extends Application {
+
   @Override
   public void start(Stage stage) throws IOException {
-    FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("dropdown.fxml"));
+    FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello.fxml"));
     Scene scene = new Scene(fxmlLoader.load(), 320, 240);
     stage.setTitle("Hello!");
     stage.setScene(scene);
