--- src/main/resources/at/fh/burgenland/dropdown.fxml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="at.fh.burgenland.HelloController">
-  <padding>
-    <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-  </padding>
-
-  <ComboBox fx:id="audioInputComboBox" layoutX="250.0" layoutY="14.0" prefWidth="150.0" promptText="(Mikrophonauswahl)" />
-  <Button fx:id="recordButton" layoutX="14.0" layoutY="14.0" mnemonicParsing="false" text="Aufnehmen" />
-
-  <Label fx:id="welcomeText" />
-  <Button onAction="#onHelloButtonClick" text="Hello!" />
-
-
-</VBox>
