+++ src/main/resources/at/fh/burgenland/hello.fxml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/17.0.12"
+  xmlns:fx="http://javafx.com/fxml/1" fx:controller="at.fh.burgenland.HelloController">
+  <padding>
+    <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
+  </padding>
+
+  <fx:include source="input_controls.fxml" fx:id="inputControlsPane"/>
+
+  <Label fx:id="welcomeText"/>
+  <Button onAction="#onHelloButtonClick" text="Hello!"/>
+
+
+</VBox>
